# BE/myapp/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/admin/events/all": {
      "get": {
        "tags": ["Admin API"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/worker/events/apply": {
      "post": {
        "tags": ["Worker API"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "workerId": {
                  "example": "any"
                },
                "userName": {
                  "example": "any"
                },
                "start": {
                  "example": "any"
                },
                "end": {
                  "example": "any"
                },
                "lastShiftStart": {
                  "example": "any"
                },
                "lastShiftEnd": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/worker/events/apply/{workerId}": {
      "get": {
        "tags": ["Worker API"],
        "description": "",
        "parameters": [
          {
            "name": "workerId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/home/signup": {
      "post": {
        "tags": ["Account API"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "userName": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "userType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/home/login": {
      "post": {
        "tags": ["Account API"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "userType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/home/logout": {
      "post": {
        "tags": ["Account API"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}